User view and Dialog role assignment
Displays the list of users in the userview box, and allows the user to change their username that gets updated to everyone after a timeout has occured. Dialog box pops whenever someone assigns the roles, giving everyone a role and it updates everytime someone clicks on the role assigner button.